nbins should be an int

+++ global_vars.h
@@ -82,7 +82,7 @@ int read_gadget_head(gadget_header *out_header, FILE *fd, int old);
 void help(void);
 /* These functions do the work*/
 int powerspectrum(const int dims, double *field, double *power);
-double mean_flux(double * tau, double nbins, double obs_flux, double tol);
+double mean_flux(double * tau, int nbins, double obs_flux, double tol);
 void calc_power_spectra(double *flux_power, double *tau_H1,double scale, double tau_eff, int NumLos);
 void calc_pdf(double *flux_pdf, double *tau_H1,double scale, int NumLos);
